Refactor run function outside of TestSet
