+++ qpsolvers_benchmark/test_set.py
@@ -20,46 +20,15 @@ Base class for test sets.
 """
 
 import abc
-from time import perf_counter
-from typing import Dict, Iterator, Optional, Tuple
+from typing import Dict, Iterator, Optional
 
 import qpsolvers
-from qpsolvers.exceptions import SolverNotFound
 
 from .exceptions import ProblemNotFound
 from .problem import Problem
-from .results import Results
-from .solver_issues import skip_solver_issue, skip_solver_timeout
 from .solver_settings import SolverSettings
 from .spdlog import logging
 from .tolerance import Tolerance
-
-
-def time_solve_problem(
-    problem, solver: str, **kwargs
-) -> Tuple[qpsolvers.Solution, float]:
-    """
-    Solve quadratic program.
-
-    Args:
-        problem: Quadratic program to solve.
-        solver: Name of the backend QP solver to call.
-
-    Returns:
-        Solution to the quadratic program, along with the time the solver took
-        to compute it.
-    """
-    # Don't time matrix conversions for solvers that require sparse inputs
-    if solver in ["highs", "osqp", "scs"]:
-        problem = problem.to_sparse()
-    start_time = perf_counter()
-    try:
-        solution = qpsolvers.solve_problem(problem, solver=solver, **kwargs)
-    except Exception as e:
-        logging.warning(f"Caught solver exception: {e}")
-        solution = qpsolvers.Solution(problem)
-    runtime = perf_counter() - start_time
-    return solution, runtime
 
 
 class TestSet(abc.ABC):
@@ -175,108 +144,3 @@ class TestSet(abc.ABC):
             f"problem '{name}' not found "
             f"in the {self.__class__.__name__} test set"
         )
-
-    def run(
-        self,
-        results: Results,
-        only_problem: Optional[str] = None,
-        only_settings: Optional[str] = None,
-        only_solver: Optional[str] = None,
-        rerun: bool = False,
-        include_timeouts: bool = False,
-    ) -> None:
-        """
-        Run test set.
-
-        Args:
-            results: Results instance to write to.
-            only_problem: If set, only run that specific problem in the set.
-            only_settings: If set, only run with these solver settings.
-            only_solver: If set, only run that specific solver.
-            rerun: If set, rerun instances that already have a result.
-            include_timeouts: If set, also rerun known timeouts.
-        """
-        if only_settings and only_settings not in self.solver_settings:
-            raise ValueError(
-                f"settings '{only_settings}' not in the list of settings "
-                f"for this test set: {list(self.solver_settings.keys())}"
-            )
-        if only_solver and only_solver not in self.solvers:
-            raise SolverNotFound(
-                f"solver '{only_solver}' not in the list of "
-                f"available solvers for this test set: {self.solvers}"
-            )
-
-        filtered_solvers = [
-            solver
-            for solver in self.solvers
-            if only_solver is None or solver == only_solver
-        ]
-        filtered_settings = [
-            settings
-            for settings in self.solver_settings
-            if only_settings is None or settings == only_settings
-        ]
-
-        nb_called = 0
-        start_counter = perf_counter()
-        for problem in self:
-            if only_problem and problem.name != only_problem:
-                continue
-            for solver in filtered_solvers:
-                for settings in filtered_settings:
-                    time_limit = self.tolerances[settings].runtime
-                    if skip_solver_issue(problem, solver):
-                        failure = (
-                            problem,
-                            solver,
-                            settings,
-                            qpsolvers.Solution(problem),
-                            0.0,
-                        )
-                        results.update(*failure)
-                        continue
-                    if skip_solver_timeout(
-                        time_limit, problem, solver, settings
-                    ):
-                        failure = (
-                            problem,
-                            solver,
-                            settings,
-                            qpsolvers.Solution(problem),
-                            0.0,
-                        )
-                        results.update(*failure)
-                        continue
-                    if results.has(problem, solver, settings):
-                        if not rerun:
-                            logging.debug(
-                                f"{problem.name} already solved by {solver} "
-                                f"with {settings} settings..."
-                            )
-                            continue
-                        if not include_timeouts and results.is_timeout(
-                            problem, solver, settings, time_limit
-                        ):
-                            logging.info(
-                                f"Skipping {problem.name} with {solver} and "
-                                f"{settings} settings as a previous timeout..."
-                            )
-                            continue
-                    logging.info(
-                        f"Solving {problem.name} by {solver} "
-                        f"with {settings} settings..."
-                    )
-                    kwargs = self.solver_settings[settings][solver]
-                    solution, runtime = time_solve_problem(
-                        problem, solver, **kwargs
-                    )
-                    results.update(
-                        problem, solver, settings, solution, runtime
-                    )
-                    results.write()
-                    nb_called += 1
-
-        duration = perf_counter() - start_counter
-        logging.info(f"Ran the test set in {duration:.0f} seconds")
-        logging.info(f"Made {nb_called} QP solver calls")
